Finished BlockListener
