+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/listener/InventoryMoveItemListener.java
@@ -1,0 +1,30 @@
+package net.countercraft.movecraft.listener;
+
+import net.countercraft.movecraft.MovecraftLocation;
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.utils.MathUtils;
+import org.bukkit.block.Hopper;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.inventory.InventoryMoveItemEvent;
+
+public class InventoryMoveItemListener implements Listener {
+    // prevent hoppers on cruising crafts
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onHopperEvent(InventoryMoveItemEvent event) {
+        if (!(event.getSource().getHolder() instanceof Hopper)) {
+            return;
+        }
+        Hopper block = (Hopper) event.getSource().getHolder();
+        CraftManager.getInstance().getCraftsInWorld(block.getWorld());
+        for (Craft tcraft : CraftManager.getInstance().getCraftsInWorld(block.getWorld())) {
+            MovecraftLocation mloc = new MovecraftLocation(block.getX(), block.getY(), block.getZ());
+            if (MathUtils.locIsNearCraftFast(tcraft, mloc) && tcraft.getCruising() && !tcraft.isNotProcessing()) {
+                event.setCancelled(true);
+                return;
+            }
+        }
+    }
+}
